Working on making model export more object oriented
This isn't done yet, just pushing to have a backup
